♻️ Consolidate benchmarking scripts
